+++ lib/justifi/payment.rb
@@ -34,6 +34,13 @@ module Justifi
           params: params,
           headers: {})
       end
+
+      def capture(payment_id:, amount:, headers: {}, idempotency_key: nil)
+        JustifiOperations.idempotently_request("/v1/payments/#{payment_id}/capture",
+          method: :post,
+          params: {amount: amount},
+          headers: {})
+      end
     end
   end
 end
